requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
sendgrid>=6.10.0
googlemaps>=4.10.0
firebase-admin>=6.2.0
django-filter>=23.0.0
requests>=2.31.0